dinput: Move controller vidpid and name wrangling to a separate file.
... so we can share it with winebus.sys/bus_sdl.c

CW-Bug-Id: #18761
