+++ dlls/dinput/joystick_sdl.c
@@ -58,103 +58,7 @@
 
 WINE_DEFAULT_DEBUG_CHANNEL(dinput);
 
-#define VID_SONY 0x054c
-#define PID_SONY_DUALSHOCK_4 0x05c4
-#define PID_SONY_DUALSHOCK_4_2 0x09cc
-#define PID_SONY_DUALSHOCK_4_DONGLE 0x0ba0
-#define PID_SONY_DUALSENSE 0x0ce6
-
-#define VID_VALVE 0x28de
-#define PID_VALVE_VIRTUAL_CONTROLLER 0x11ff
-
-#define VID_MICROSOFT 0x045e
-#define PID_MICROSOFT_XBOX_360 0x028e
-#define PID_MICROSOFT_XBOX_360_WIRELESS 0x028f
-#define PID_MICROSOFT_XBOX_360_ADAPTER  0x0719
-#define PID_MICROSOFT_XBOX_ONE 0x02d1
-#define PID_MICROSOFT_XBOX_ONE_CF 0x02dd
-#define PID_MICROSOFT_XBOX_ONE_ELITE 0x02e3
-#define PID_MICROSOFT_XBOX_ONE_S 0x02ea
-#define PID_MICROSOFT_XBOX_ONE_S_2 0x02fd
-
-static BOOL __controller_hack_sdl_is_vid_pid_xbox_360(WORD vid, WORD pid)
-{
-    if(vid != VID_MICROSOFT)
-        return FALSE;
-
-    if(pid == PID_MICROSOFT_XBOX_360 ||
-        pid == PID_MICROSOFT_XBOX_360_WIRELESS ||
-        pid == PID_MICROSOFT_XBOX_360_ADAPTER)
-        return TRUE;
-
-    return FALSE;
-}
-static BOOL __controller_hack_sdl_is_vid_pid_xbox_one(WORD vid, WORD pid) {
-    if(vid != VID_MICROSOFT)
-        return FALSE;
-
-    if(pid == PID_MICROSOFT_XBOX_ONE ||
-        pid == PID_MICROSOFT_XBOX_ONE_CF ||
-        pid == PID_MICROSOFT_XBOX_ONE_ELITE ||
-        pid == PID_MICROSOFT_XBOX_ONE_S ||
-        pid == PID_MICROSOFT_XBOX_ONE_S_2)
-        return TRUE;
-
-    return FALSE;
-}
-
-static BOOL __controller_hack_sdl_is_vid_pid_sony_dualshock4(WORD vid, WORD pid)
-{
-    if(vid != VID_SONY)
-        return FALSE;
-
-    if(pid == PID_SONY_DUALSHOCK_4 ||
-        pid == PID_SONY_DUALSHOCK_4_2 ||
-        pid == PID_SONY_DUALSHOCK_4_DONGLE)
-        return TRUE;
-
-    return FALSE;
-}
-
-static BOOL __controller_hack_sdl_is_vid_pid_sony_dualsense(WORD vid, WORD pid)
-{
-    if(vid == VID_SONY && pid == PID_SONY_DUALSENSE)
-        return TRUE;
-
-    return FALSE;
-}
-
-static void __controller_hack_sdl_vid_pid_override(WORD *vid, WORD *pid)
-{
-    TRACE("(*vid = %04hx, *pid = *%04hx\n", *vid, *pid);
-    if(*vid == VID_VALVE && *pid == PID_VALVE_VIRTUAL_CONTROLLER)
-    {
-        TRACE("Valve Virtual Controller found - pretending it's x360\n");
-        *vid = VID_MICROSOFT;
-        *pid = PID_MICROSOFT_XBOX_360;
-    }
-}
-
-static void __controller_hack_sdl_name_override(WORD vid, WORD pid, const char **name)
-{
-    const char *new_name = NULL;
-
-    TRACE("vid = %04hx, pid = %04hx, *name = %s\n", vid, pid, debugstr_a(*name));
-
-    if(__controller_hack_sdl_is_vid_pid_xbox_360(vid, pid))
-        new_name = "Controller (XBOX 360 For Windows)";
-    else if(__controller_hack_sdl_is_vid_pid_xbox_one(vid, pid))
-        new_name = "Controller (XBOX One For Windows)";
-    else if(__controller_hack_sdl_is_vid_pid_sony_dualshock4(vid, pid) ||
-             __controller_hack_sdl_is_vid_pid_sony_dualsense(vid, pid))
-        new_name = "Wireless Controller";
-
-    if(new_name)
-    {
-        TRACE("new name for the controller = %s\n", debugstr_a(new_name));
-        *name = new_name;
-    }
-}
+#include "wine/sdl_controller_overrides.h"
 
 typedef struct JoystickImpl JoystickImpl;
 static const IDirectInputDevice8AVtbl JoystickAvt;
